split challenge.go
